fix(docs): move Other Projects sidebar to server-side loading for SEO
Move the "Other Projects" sidebar section from client-side onMount fetch
to server-side data loading via +layout.server.ts so that backlinks are
present in the initial HTML for crawlers (Google, Ahrefs, etc.).

- Add +layout.server.ts that fetches /api/other-projects at request time
- Pass otherProjects data from layout to Sidebar as a prop
- Remove client-side onMount fetch, OtherProject type, and formatTitle
  helper from Sidebar.svelte
